Keep instance reclaim in the Yaegi root.
The plugin package constructs and serves; Traefik New owns the process-table reuse.
