+++ knowledge/devdocs/core_geoblock_plugin_instance.md
@@ -8,35 +8,39 @@ _Avoid_: treating each returned `*Plugin` as a distinct core (it is a shallow co
 
 ## Overview
 
-`New` builds the Plugin once per name+config and reuses it for later routers and same-config reloads. Each `New` still returns a `*Plugin` whose `next` is that router’s chain and whose provider was opened with that `New` context.
+Root `New` builds the Plugin once per name+config and reuses it for later routers and same-config reloads. Each `New` still returns a `*geoblock.Plugin` whose `next` is that router’s chain and whose provider was opened with that `New` context.
 
 ## How to use
 
-- Do not construct maps, regexes, IP helpers, or ban HTML outside `newPluginCore`. That create runs once per incarnation.
+- Do not construct maps, regexes, IP helpers, or ban HTML outside `geoblock.NewCore`. That create runs once per incarnation.
 - Always open the DatabaseProvider with this `New` context before `reclaim.Open`, so wrappers bind even on reuse.
 - Key prefix is `plugin:`. Same process table as `bin:` / `mmdb:` (`std_go_reclaim.md`).
 - Dispose is empty. After grace the table drops the incarnation.
-- Tests: `dbwrappers.Reset` / `ResetWith`. Assert shared pointers (`allowedIPBlocks`, country maps) and per-`New` `next`. Filter reclaim logs with the `plugin:` prefix.
+- Tests: `dbwrappers.Reset` / `ResetWith` from the root package. Assert `SameCore` and per-`New` `Next`. Filter reclaim logs with the `plugin:` prefix.
 
 ## Pattern snippet
 
 ```go
+if err := geoblock.Prepare(cfg, name, logger); err != nil {
+	return nil, err
+}
+db, err := geoblock.OpenDatabase(ctx, cfg, bootstrapLogger)
 v, err := reclaim.Open(ctx, pluginKey(name, cfg), func() (any, error) {
-	return newPluginCore(name, cfg, db, logger, enrich)
+	return geoblock.NewCore(name, cfg, db, logger)
 }, func(any) {})
-out := *(v.(*Plugin))
-out.next = next
-out.db = db
+return v.(*geoblock.Plugin).ForRoute(next, db), nil
 ```
 
 ## Key files
 
 - `plugin.go` — `New`, `bindPlugin`, `pluginKey`
-- `plugin_lifecycle_test.go` — share, miss, reclaim, grace
+- `plugin_instance_test.go` — share, miss, reclaim, grace
+- `pkg/geoblock` — Prepare, OpenDatabase, NewCore, ForRoute
 - `pkg/reclaim` — process table
 
 ## Gotchas
 
-- Hash is JSON+FNV of `Config` after defaults, pointer fallbacks, empty-source binds, temp auto-update dir, and ban-HTML path search.
+- Hash is JSON+FNV of `Config` after `Prepare` (defaults, pointer fallbacks, empty-source binds, temp auto-update dir, ban-HTML path search).
 - Two middleware names never share, even with the same config.
 - Do not write `Table[*Plugin]` (Yaegi).
+- `pkg/geoblock.New` constructs one Plugin and does not reuse. Instance tests must call root `New`.
