+++ pkg/geoblock/plugin_ipheaders_test.go
@@ -1,4 +1,4 @@
-package traefik_geoblock
+package geoblock
 
 import (
 	"context"
