+++ pkg/geoblock/plugin_observe_test.go
@@ -1,4 +1,4 @@
-package traefik_geoblock
+package geoblock
 
 import (
 	"context"
